+++ ERPSystem/ERP.FournisseurService/Program.cs
@@ -1,3 +1,4 @@
+using Confluent.Kafka;
 using ERP.FournisseurService.Application.Interfaces;
 using ERP.FournisseurService.Application.Services;
 using ERP.FournisseurService.Infrastructure.Messaging;
@@ -5,8 +6,10 @@ using ERP.FournisseurService.Infrastructure.Persistence;
 using ERP.FournisseurService.Infrastructure.Persistence.Repositories;
 using ERP.FournisseurService.Infrastructure.Persistence.Seeders;
 using ERP.FournisseurService.Middleware;
+using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
 using System.Text.Json.Serialization;
 
 WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
@@ -22,6 +25,19 @@ string connectionString = builder.Configuration.GetConnectionString("DefaultConn
 builder.Services.AddDbContext<FournisseurDbContext>(options =>
     options.UseSqlServer(connectionString));
 
+
+var kafkaConfig = new ProducerConfig
+{
+    BootstrapServers = builder.Configuration["Kafka:BootstrapServers"]
+};
+
+///////////////////////////////////////////////////
+// Health Checks
+///////////////////////////////////////////////////
+builder.Services.AddHealthChecks()
+    .AddSqlServer(connectionString, name: "sql")
+    .AddKafka(kafkaConfig)
+    .AddCheck("self", () => HealthCheckResult.Healthy());
 // =========================
 // DEPENDENCY INJECTION
 // =========================
@@ -78,7 +94,9 @@ using (IServiceScope scope = app.Services.CreateScope())
     FournisseurDbContext db = scope.ServiceProvider.GetRequiredService<FournisseurDbContext>();
     DatabaseSeeder seeder = scope.ServiceProvider.GetRequiredService<DatabaseSeeder>();
 
-    await db.Database.EnsureDeletedAsync();
+    if (app.Environment.IsDevelopment())
+        await db.Database.EnsureDeletedAsync();
+
     await db.Database.MigrateAsync();
     await seeder.SeedAsync();
 }
@@ -86,8 +104,11 @@ using (IServiceScope scope = app.Services.CreateScope())
 // =========================
 // PIPELINE
 // =========================
-app.UseSwagger();
-app.UseSwaggerUI();
+if(app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
 
 if (!app.Environment.IsDevelopment())
 {
@@ -97,5 +118,15 @@ if (!app.Environment.IsDevelopment())
 
 app.UseMiddleware<GlobalExceptionMiddleware>();
 app.MapControllers();
+
+app.MapHealthChecks("/health");
+app.MapHealthChecks("/health/live", new HealthCheckOptions
+{
+    Predicate = r => r.Name == "self"
+});
+app.MapHealthChecks("/health/ready", new HealthCheckOptions
+{
+    Predicate = r => r.Name is "sql" or "kafka"
+});
 
 app.Run();
